api/requirements.txt
psycopg2-binary
redis>=2.10.6
flask>=0.12